Removed GetPeers from the client task, reduce cloning
The GetPeers service request is resolved by using the topology object
available to the network task. No need for intercom roundtrips,
also all use of the topology object is removed from the client task.
